Move loader stuff into `loader` module in `bevy_gltf`
