--- crates/bevy_gltf/src/helper_types.rs
@@ -1,137 +1,0 @@
-use std::path::PathBuf;
-
-use bevy_asset::{Handle, LoadContext};
-use bevy_ecs::{entity::Entity, name::Name};
-use bevy_image::{Image, ImageLoaderSettings, ImageSampler, ImageSamplerDescriptor};
-use bevy_math::Vec3;
-use bevy_render::mesh::morph::MorphAttributes;
-use gltf::accessor::Iter;
-use serde::Deserialize;
-use smallvec::SmallVec;
-
-use crate::GltfAssetLabel;
-
-pub struct DataUri<'a> {
-    pub mime_type: &'a str,
-    pub base64: bool,
-    pub data: &'a str,
-}
-
-fn split_once(input: &str, delimiter: char) -> Option<(&str, &str)> {
-    let mut iter = input.splitn(2, delimiter);
-    Some((iter.next()?, iter.next()?))
-}
-
-impl<'a> DataUri<'a> {
-    pub fn parse(uri: &'a str) -> Result<DataUri<'a>, ()> {
-        let uri = uri.strip_prefix("data:").ok_or(())?;
-        let (mime_type, data) = split_once(uri, ',').ok_or(())?;
-
-        let (mime_type, base64) = match mime_type.strip_suffix(";base64") {
-            Some(mime_type) => (mime_type, true),
-            None => (mime_type, false),
-        };
-
-        Ok(DataUri {
-            mime_type,
-            base64,
-            data,
-        })
-    }
-
-    pub fn decode(&self) -> Result<Vec<u8>, base64::DecodeError> {
-        if self.base64 {
-            base64::Engine::decode(&base64::engine::general_purpose::STANDARD, self.data)
-        } else {
-            Ok(self.data.as_bytes().to_owned())
-        }
-    }
-}
-
-pub enum ImageOrPath {
-    Image {
-        image: Image,
-        label: GltfAssetLabel,
-    },
-    Path {
-        path: PathBuf,
-        is_srgb: bool,
-        sampler_descriptor: ImageSamplerDescriptor,
-    },
-}
-
-impl ImageOrPath {
-    // TODO: use the threaded impl on wasm once wasm thread pool doesn't deadlock on it
-    // See https://github.com/bevyengine/bevy/issues/1924 for more details
-    // The taskpool use is also avoided when there is only one texture for performance reasons and
-    // to avoid https://github.com/bevyengine/bevy/pull/2725
-    // PERF: could this be a Vec instead? Are gltf texture indices dense?
-    pub fn process_loaded_texture(
-        self,
-        load_context: &mut LoadContext,
-        handles: &mut Vec<Handle<Image>>,
-    ) {
-        let handle = match self {
-            ImageOrPath::Image { label, image } => {
-                load_context.add_labeled_asset(label.to_string(), image)
-            }
-            ImageOrPath::Path {
-                path,
-                is_srgb,
-                sampler_descriptor,
-            } => load_context
-                .loader()
-                .with_settings(move |settings: &mut ImageLoaderSettings| {
-                    settings.is_srgb = is_srgb;
-                    settings.sampler = ImageSampler::Descriptor(sampler_descriptor.clone());
-                })
-                .load(path),
-        };
-        handles.push(handle);
-    }
-}
-
-pub(super) struct PrimitiveMorphAttributesIter<'s>(
-    pub  (
-        Option<Iter<'s, [f32; 3]>>,
-        Option<Iter<'s, [f32; 3]>>,
-        Option<Iter<'s, [f32; 3]>>,
-    ),
-);
-
-impl<'s> Iterator for PrimitiveMorphAttributesIter<'s> {
-    type Item = MorphAttributes;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let position = self.0 .0.as_mut().and_then(Iterator::next);
-        let normal = self.0 .1.as_mut().and_then(Iterator::next);
-        let tangent = self.0 .2.as_mut().and_then(Iterator::next);
-        if position.is_none() && normal.is_none() && tangent.is_none() {
-            return None;
-        }
-
-        Some(MorphAttributes {
-            position: position.map(Into::into).unwrap_or(Vec3::ZERO),
-            normal: normal.map(Into::into).unwrap_or(Vec3::ZERO),
-            tangent: tangent.map(Into::into).unwrap_or(Vec3::ZERO),
-        })
-    }
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MorphTargetNames {
-    pub target_names: Vec<String>,
-}
-
-// A helper structure for `load_node` that contains information about the
-// nearest ancestor animation root.
-#[cfg(feature = "bevy_animation")]
-#[derive(Clone)]
-pub struct AnimationContext {
-    // The nearest ancestor animation root.
-    pub root: Entity,
-    // The path to the animation root. This is used for constructing the
-    // animation target UUIDs.
-    pub path: SmallVec<[Name; 8]>,
-}
