+++ crates/bevy_gltf/src/lib.rs
@@ -92,7 +92,6 @@
 
 mod assets;
 mod ext;
-mod helper_types;
 mod label;
 mod loader;
 mod vertex_attributes;
