+++ crates/bevy_gltf/src/loader/data_uri.rs
@@ -1,0 +1,36 @@
+pub struct DataUri<'a> {
+    pub mime_type: &'a str,
+    pub base64: bool,
+    pub data: &'a str,
+}
+
+fn split_once(input: &str, delimiter: char) -> Option<(&str, &str)> {
+    let mut iter = input.splitn(2, delimiter);
+    Some((iter.next()?, iter.next()?))
+}
+
+impl<'a> DataUri<'a> {
+    pub fn parse(uri: &'a str) -> Result<DataUri<'a>, ()> {
+        let uri = uri.strip_prefix("data:").ok_or(())?;
+        let (mime_type, data) = split_once(uri, ',').ok_or(())?;
+
+        let (mime_type, base64) = match mime_type.strip_suffix(";base64") {
+            Some(mime_type) => (mime_type, true),
+            None => (mime_type, false),
+        };
+
+        Ok(DataUri {
+            mime_type,
+            base64,
+            data,
+        })
+    }
+
+    pub fn decode(&self) -> Result<Vec<u8>, base64::DecodeError> {
+        if self.base64 {
+            base64::Engine::decode(&base64::engine::general_purpose::STANDARD, self.data)
+        } else {
+            Ok(self.data.as_bytes().to_owned())
+        }
+    }
+}
