+++ crates/bevy_gltf/src/loader/morph_target_names.rs
@@ -1,0 +1,7 @@
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct MorphTargetNames {
+    pub target_names: Vec<String>,
+}
